// generated/src/app/api-models/aptax-report-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AptaxReportSetup {
  seqAptaxId: number;
  jobId: string;
  requestUser: string;
  requestDatetime: Date;
  requestType: string;
  reportType: string;
  action: string;
  status: string;
  taxRepEntity: string;
  taxYear: Date;
  dollarLimit: number;
  comments: string;
  sortOptions: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  template: string;
  inProcess: string;
  includeNochecks: string;
  includeOffsets: string;
}